pyproject.toml
[project]
name = "review-analysis-agent"
version = "0.1.0"
description = "Agent service for reviewing PRD/TRD/testcase documents and chat"
readme = "speckit.specify.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi",
  "python-multipart",
  "uvicorn",
  "solara",
  "pydantic>=2",
  "pyyaml",
  "gitpython",
  "langchain",
  "langchain-openai",
  "alibabacloud_docmind_api20220711",
  "alibabacloud_tea_openapi",
  "alibabacloud_tea_util",
  "alibabacloud_credentials",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-cov",
  "httpx",
  "ruff",
  "mypy",
  "types-PyYAML",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["B008"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
disallow_any_generics = true
disallow_subclassing_any = true

[[tool.mypy.overrides]]
module = ["src.cli.*"]
ignore_errors = true

[[tool.mypy.overrides]]
module = ["src.utils.aili_doc_parser"]
ignore_errors = true

[tool.pytest.ini_options]
addopts = "-q --import-mode=prepend --cov=src --cov-report=term-missing --cov-fail-under=70"
pythonpath = ["."]
testpaths = ["tests"]

[tool.coverage.run]
omit = [
  "src/cli/app.py",
]
